+++ src/usuario/repositories/usuario.repository.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { ConflictException, Injectable } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { CreateUsuarioDto } from '../dto/create-usuario.dto';
 import { UpdateUsuarioDto } from '../dto/update-usuario.dto';
@@ -9,6 +9,15 @@ export class UsuarioRepository {
   constructor(private readonly prisma: PrismaService) {}
 
   async create(createUsuarioDto: CreateUsuarioDto): Promise<UsuarioEntity> {
+    const usuarioExiste = await this.prisma.usuario.findFirst({
+      where: {
+        usuario: createUsuarioDto.usuario,
+      },
+    });
+
+    if (usuarioExiste) {
+      throw new ConflictException('Usuário já existe!!');
+    }
     return this.prisma.usuario.create({
       data: createUsuarioDto,
     });
@@ -30,6 +39,15 @@ export class UsuarioRepository {
     id: string,
     updateUsuarioDto: UpdateUsuarioDto,
   ): Promise<UsuarioEntity> {
+    const usuarioExiste = await this.prisma.usuario.findFirst({
+      where: {
+        id,
+      },
+    });
+
+    if (!usuarioExiste) {
+      throw new ConflictException('Usuário não existe!!');
+    }
     return this.prisma.usuario.update({
       where: {
         id,
@@ -39,6 +57,15 @@ export class UsuarioRepository {
   }
 
   async remove(id: string): Promise<UsuarioEntity> {
+    const usuarioExiste = await this.prisma.usuario.findFirst({
+      where: {
+        id,
+      },
+    });
+
+    if (!usuarioExiste) {
+      throw new ConflictException('Esse usuario não existe!!');
+    }
     return this.prisma.usuario.delete({
       where: {
         id,
